refactor: move bans to server/bans and modernize endpoints
Reorganized ban endpoints to match server endpoint structure:
- Moved routers/v2/bans to routers/v2/server/bans
- Renamed bans.py to endpoints.py, ban_models.py to models.py
- Updated to use linkd dependency injection
- Updated to use utils.security.check_authentication
- Changed routes from /v2/ban/* to /v2/server/{server_id}/bans
- Added count field to list response
- Updated imports in startup.py

New endpoints:
- GET /v2/server/{server_id}/bans (was /v2/ban/list/{server_id})
- POST /v2/server/{server_id}/bans/{player_tag} (was /v2/ban/add/...)
- DELETE /v2/server/{server_id}/bans/{player_tag} (was /v2/ban/remove/...)

This provides a consistent API structure with strikes, links, roles, etc.
