+++ startup.py
@@ -18,6 +18,7 @@ from routers.v2.server.links.endpoints import router as server_links_router
 from routers.v2.server.clans.endpoints import router as server_clans_router
 from routers.v2.server.roles.endpoints import router as server_roles_router
 from routers.v2.server.strikes.endpoints import router as server_strikes_router
+from routers.v2.server.bans.endpoints import router as server_bans_router
 from routers.v2.capital.capital import router as capital_router
 from routers.v2.activity.activity import router as activity_router
 from routers.v2.legends.legends import router as legends_router
@@ -45,6 +46,7 @@ def define_app(app: FastAPI):
     app.include_router(server_clans_router)
     app.include_router(server_roles_router)
     app.include_router(server_strikes_router)
+    app.include_router(server_bans_router)
     app.include_router(capital_router)
     app.include_router(activity_router)
     app.include_router(legends_router)
